refactor(ai): rename action-context to tool-bridge and expand wire payload
Rename ServerToolDefinition → ToolDefinitionPayload and toServerDefinitions →
toToolDefinitions to avoid confusion with TanStack AI's ServerTool concept.

Forward outputSchema, needsApproval, and metadata in the wire payload—without
needsApproval the server's approval flow silently auto-executes mutations.
